uso de dto en obtener por inscriptos

+++ integrador-3/src/main/java/org/grupo14/integrador3/dto/CarreraDTO.java
@@ -8,8 +8,8 @@ public class CarreraDTO {
     private String nombre;
     private int cantidad;
 
-    public CarreraDTO(String carrera, int cantidad) {
+    public CarreraDTO(String carrera, Long cantidad) {
           this.nombre = carrera;
-          this.cantidad = cantidad;
+          this.cantidad = Math.toIntExact(cantidad);
     }
 }
